It builds. It's unlikely that this code will actually work yet because of: 	- Lifetime management and sharing issues previously hidden by reliance on 	  PdfVecObjects for things that don't need to be indirect objects; 	- Areas where the code assumes the object will be created indirect and relies on that, 	  but we now create a direct object by default 	- Areas where the code assumes it's been passed an indirect object when it might not have 	  (this is quite common) ... and there's probably more. This is not small stuff, that's why I'm trialling it in a branch.
I expect it'll be necessary to implement more carefully managed object lifetimes and some form of safe
sharing now that PdfVecObjects isn't the central owner/manager of all objects. That method might well be
"make PdfVecObjects the central owner/manager again" - we'll see.

I'm not even certain that using PdfVariant as the base "indirect object" class is best. We might be better off
with PdfObject representing any PDF object (direct or indirect), and PdfIObject or something for indirect objects.
That'd leave variants as the simple data container for all of those, which would be nice - but then, PdfObject would
basically be a wrapper class around PdfVariant. Only really useful if we land up doing some kind of internal sharing
mechanism for lifetime management.

Anyhow, here's progress to date.



git-svn-id: http://svn.code.sf.net/p/podofo/code/podofo/branches/VarObjCleanup@498 e6e4d1b7-ffb3-4f57-9d89-6d52b3892af5