Fix heap-buffer-overflow in PdfString::ConvertUTF16toUTF8
I analyzed this based on a report to the podofo-users mailing list
(metadata below) and a PoC file attached to a post with the same (after
the brackets) subject to the podofo-svn mailing list. This is tested
with the PoC and podofo-test compiled with GCC 4.8.4, clang 3.8.1-12~bpo8+1
and on a newer system GCC 7.3.0 and clang 7.0.0.3~ubuntu0.18.04.1 also with
ASan (all OK).

Post-Subject: [Podofo-users] Seems bug in OutputInfoDict() in pdfinfo.cpp
Post-Approved-Date: Wed, 23 Jan 2019 21:15:05 +0000
Post-Message-ID: <CAD3bFv0VMb1DHK0wta0t=Xg45Oc0gW+iS1dYV-Cxpk7hKBoeZQ@mail.gmail.com>
Post-Archive-Date: Date: Wed, 23 Jan 2019 19:53:59 +0000
Post-Archive-URL: https://sourceforge.net/p/podofo/mailman/message/36525641/ 


git-svn-id: http://svn.code.sf.net/p/podofo/code/podofo/trunk@1969 e6e4d1b7-ffb3-4f57-9d89-6d52b3892af5